HeadCollaboration/requirements.txt
fairseq.egg==info
kmeans_pytorch.egg==info
omegaconf==2.0.6
pyserini==0.14.0
scikit_learn==0.23.2
torch==1.9.0+cu111
python==3.8.5
wandb==0.12.9 